daily

+++ FoodDiary/FoodDiary/Models/Foods.cs
@@ -10,9 +10,9 @@ namespace FoodDiary.Models
         [PrimaryKey, AutoIncrement]
         public int ID { get; set; }
         public string Name { get; set; }
-        public int Calory { get; set; }
-        public int Carbohydrates { get; set; }
-        public int Fats { get; set; }
-        public int Proteins { get; set; }
+        public double Calory { get; set; }
+        public double Carbohydrates { get; set; }
+        public double Fats { get; set; }
+        public double Proteins { get; set; }
     }
 }
